fix for gateway
gateway記載に際して、serviceが複数だったところを他のAppsと合わせるため単一に変更
