+++ docker-compose.yml
@@ -50,7 +50,7 @@ services:
       - AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID}
       - AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY}
       - AWS_SESSION_TOKEN=${AWS_SESSION_TOKEN}
-      - SQS_ENDPOINT=${ENDPOINT}
+      - SQS_ENDPOINT=${SQS_ENDPOINT}
       - DIAGNOSIS_JIRA_QUEUE_URL=${DIAGNOSIS_JIRA_QUEUE_URL}
       - DIAGNOSIS_JIRA_QUEUE_NAME=${DIAGNOSIS_JIRA_QUEUE_NAME}
       - DIAGNOSIS_JIRA_URL=${DIAGNOSIS_JIRA_URL}
