+++ env_sample.sh
@@ -20,7 +20,7 @@ export DB_LOG_MODE="false"
 export AWS_REGION="ap-northeast-1"
 export AWS_ACCESS_KEY_ID="testkey"
 export AWS_SECRET_ACCESS_KEY="testsecretkey"
-export ENDPOINT="http://sqs:9324"
+export SQS_ENDPOINT="http://sqs:9324"
 export DIAGNOSIS_JIRA_QUEUE_NAME="diagnosis-jira"
 export DIAGNOSIS_JIRA_QUEUE_URL="http://sqs:9324/queue/diagnosis-jira"
 export DIAGNOSIS_JIRA_URL="https://ca-security.atlassian.net/"
