+++ pkg/pb/diagnosis/go.mod
@@ -3,6 +3,59 @@ module github.com/CyberAgent/mimosa-diagnosis/pkg/pb/diagnosis
 go 1.14
 
 require (
+	fyne.io/fyne v1.3.3 // indirect
+	gioui.org v0.0.0-20200829162755-829ee4559c5a // indirect
+	github.com/BurntSushi/xgb v0.0.0-20200324125942-20f126ea2843 // indirect
+	github.com/aclements/go-gg v0.0.0-20170323211221-abd1f791f5ee // indirect
+	github.com/aclements/go-moremath v0.0.0-20190830160640-d16893ddf098 // indirect
+	github.com/ajstarks/deck v0.0.0-20200831202436-30c9fc6549a9 // indirect
+	github.com/ajstarks/deck/generate v0.0.0-20200831202436-30c9fc6549a9 // indirect
+	github.com/ajstarks/fc v0.0.0-20200729133317-f1729ab2efcb // indirect
+	github.com/ajstarks/giocanvas v0.0.0-20200831201938-f9a05979ddb3 // indirect
+	github.com/ajstarks/svgo v0.0.0-20200725142600-7a3c8b57fecb // indirect
+	github.com/bradfitz/gomemcache v0.0.0-20190913173617-a41fca850d0b // indirect
+	github.com/coreos/pkg v0.0.0-20180928190104-399ea9e2e55f // indirect
+	github.com/cznic/cc v0.0.0-20181122101902-d673e9b70d4d // indirect
+	github.com/cznic/fileutil v0.0.0-20181122101858-4d67cfea8c87 // indirect
+	github.com/cznic/golex v0.0.0-20181122101858-9c343928389c // indirect
+	github.com/cznic/internal v0.0.0-20181122101858-3279554c546e // indirect
+	github.com/cznic/ir v0.0.0-20181122101859-da7ba2ecce8b // indirect
+	github.com/cznic/lex v0.0.0-20181122101858-ce0fb5e9bb1b // indirect
+	github.com/cznic/lexer v0.0.0-20181122101858-e884d4bd112e // indirect
+	github.com/cznic/mathutil v0.0.0-20181122101859-297441e03548 // indirect
+	github.com/cznic/strutil v0.0.0-20181122101858-275e90344537 // indirect
+	github.com/cznic/xc v0.0.0-20181122101856-45b06973881e // indirect
+	github.com/disintegration/gift v1.2.1 // indirect
+	github.com/edsrzf/mmap-go v1.0.0 // indirect
+	github.com/fogleman/gg v1.3.0 // indirect
+	github.com/garyburd/redigo v1.6.2 // indirect
+	github.com/go-ozzo/ozzo-validation/v4 v4.2.2
+	github.com/gobwas/glob v0.2.3 // indirect
+	github.com/godbus/dbus v4.1.0+incompatible // indirect
+	github.com/gogo/protobuf v1.3.1 // indirect
+	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
 	github.com/golang/protobuf v1.4.2
-	google.golang.org/grpc v1.31.0
+	github.com/gonum/blas v0.0.0-20181208220705-f22b278b28ac // indirect
+	github.com/gonum/floats v0.0.0-20181209220543-c233463c7e82 // indirect
+	github.com/gonum/internal v0.0.0-20181124074243-f884aa714029 // indirect
+	github.com/gonum/lapack v0.0.0-20181123203213-e4cdc5a0bff9 // indirect
+	github.com/gonum/matrix v0.0.0-20181209220409-c518dec07be9 // indirect
+	github.com/googleapis/gax-go v1.0.3 // indirect
+	github.com/grpc-ecosystem/grpc-gateway v1.14.7 // indirect
+	github.com/iancoleman/strcase v0.1.1 // indirect
+	github.com/jung-kurt/gofpdf v1.16.2 // indirect
+	github.com/lyft/protoc-gen-star v0.5.1 // indirect
+	github.com/mattn/go-sqlite3 v1.14.2 // indirect
+	github.com/nbutton23/zxcvbn-go v0.0.0-20180912185939-ae427f1e4c1d // indirect
+	github.com/peterbourgon/diskv v2.0.1+incompatible // indirect
+	github.com/remyoudompheng/bigfft v0.0.0-20200410134404-eec4a21b6bb0 // indirect
+	github.com/rwcarlsen/goexif v0.0.0-20190401172101-9e8deecbddbd // indirect
+	github.com/syndtr/goleveldb v1.0.0 // indirect
+	golang.org/x/build v0.0.0-20200901204317-301a72cb1ab6 // indirect
+	golang.org/x/image v0.0.0-20200801110659-972c09e46d76 // indirect
+	golang.org/x/mobile v0.0.0-20200801112145-973feb4309de // indirect
+	golang.org/x/time v0.0.0-20200630173020-3af7569d3a1e // indirect
+	google.golang.org/api v0.31.0 // indirect
+	google.golang.org/grpc v1.31.1
+	gopkg.in/yaml.v1 v1.0.0-20140924161607-9f9df34309c0 // indirect
 )
